Modernize Thrust examples following PR #753 patterns
- Remove legacy include/host_device.h headers from 40 example files
- Replace manual element assignment with std::initializer_list
- Use range-based for loops where appropriate
- Apply STL algorithms (std::generate) with lambdas
- Use .size() instead of hardcoded array sizes
- Improve semantic naming and inline usage
- Maintain compatibility with current CUDA/Thrust version
- Avoid thrust::enumerate (not available in current version)

Files modernized:
arbitrary_transformation.cu, basic_vector.cu, bounding_box.cu,
bucket_sort2d.cu, constant_iterator.cu, counting_iterator.cu,
device_ptr.cu, discrete_voronoi.cu, dot_products_with_zip.cu,
expand.cu, histogram.cu, lambda.cu, lexicographical_sort.cu,
max_abs_diff.cu, minmax.cu, mode.cu, monte_carlo.cu,
monte_carlo_disjoint_sequences.cu, norm.cu, padded_grid_reduction.cu,
permutation_iterator.cu, raw_reference_cast.cu, remove_points2d.cu,
repeated_range.cu, saxpy.cu, scan_matrix_by_rows.cu,
simple_moving_average.cu, sort.cu, sorting_aos_vs_soa.cu,
stream_compaction.cu, sum_rows.cu, summary_statistics.cu,
summed_area_table.cu, tiled_range.cu, transform_input_output_iterator.cu,
transform_iterator.cu, transform_output_iterator.cu,
uninitialized_vector.cu, weld_vertices.cu, word_count.cu

+++ thrust/examples/padded_grid_reduction.cu
@@ -12,8 +12,6 @@
 
 #include <float.h>
 
-#include "include/host_device.h"
-
 // This example computes the minimum and maximum values
 // over a padded grid.  The padded values are not considered
 // during the reduction operation.
@@ -82,9 +80,9 @@ int main()
   thrust::device_vector<float> data(M * N, -1);
 
   // initialize valid values in grid
-  for (int i = 0; i < M; i++)
+  for (size_t i = 0; i < static_cast<size_t>(M); i++)
   {
-    for (int j = 0; j < n; j++)
+    for (size_t j = 0; j < static_cast<size_t>(n); j++)
     {
       data[i * N + j] = dist(rng);
     }
@@ -93,10 +91,10 @@ int main()
   // print full grid
   std::cout << "padded grid" << std::endl;
   std::cout << std::fixed << std::setprecision(4);
-  for (int i = 0; i < M; i++)
+  for (size_t i = 0; i < static_cast<size_t>(M); i++)
   {
     std::cout << " ";
-    for (int j = 0; j < N; j++)
+    for (size_t j = 0; j < static_cast<size_t>(N); j++)
     {
       std::cout << data[i * N + j] << " ";
     }
